+++ movility/admin.py
@@ -2,7 +2,6 @@ from django.contrib import admin
 from .models import *
 
 # Register your models here.
-admin.site.register(Usuarios)
 admin.site.register(Vehiculos)
 admin.site.register(Solicitudes)
 admin.site.register(Rutas)
