+++ ubrebackend/app/profile/models.py
@@ -4,18 +4,16 @@ from ubrebackend.app.user.models import User
 
 
 class UserProfile(models.Model):
-
-    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
+    dni = models.CharField(primary_key=True, max_length=9)
+    #id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     user = models.OneToOneField(User, on_delete=models.CASCADE, related_name='profile')
-    first_name = models.CharField(max_length=50, unique=False)
+    name = models.CharField(max_length=50, unique=False)
     last_name = models.CharField(max_length=50, unique=False)
-    phone_number = models.CharField(max_length=10, unique=True, null=False, blank=False)
+    tlf = models.CharField(max_length=10, unique=True, null=False, blank=False)
     age = models.PositiveIntegerField(null=False, blank=False)
-    GENDER_CHOICES = (
-        ('M', 'Male'),
-        ('F', 'Female'),
-    )
-    gender = models.CharField(max_length=1, choices=GENDER_CHOICES)
+
+    def __str__(self):
+        return '%s %s %s' % (self.dni, self.name, self.last_name)
 
     class Meta:
         '''
