+++ ubrebackend/app/profile/views.py
@@ -15,19 +15,8 @@ class UserProfileView(RetrieveAPIView):
     def get(self, request):
         try:
             user_profile = UserProfile.objects.get(user=request.user)
-            status_code = status.HTTP_200_OK
-            response = {
-                'success': 'true',
-                'status code': status_code,
-                'message': 'User profile fetched successfully',
-                'data': [{
-                    'first_name': user_profile.first_name,
-                    'last_name': user_profile.last_name,
-                    'phone_number': user_profile.phone_number,
-                    'age': user_profile.age,
-                    'gender': user_profile.gender,
-                    }]
-                }
+            serializer = UserSerializer(user_profile)
+            return Response(serializer.data)
 
         except Exception as e:
             status_code = status.HTTP_400_BAD_REQUEST
